backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
ta==0.11.0
httpx==0.27.2
pydantic==2.10.3
python-multipart==0.0.20
yfinance==0.2.50
requests-cache==1.2.1
